Advance deprecation level of FP to lesser than Int types to ERROR #KT-30360
(cherry picked from commit 968099fbec46472b0b18809698384ba1f4b16c33)

+++ compiler/testData/codegen/bytecodeText/boxingOptimization/casts.kt
@@ -9,7 +9,7 @@ fun bar() {
     foo(1L) { x -> x!!.toByte() }
     foo(1L) { x -> x!!.toShort() }
     foo('a') { x -> x!!.toDouble() }
-    foo(1.0) { x -> x!!.toByte() }
+    foo(1.0) { x -> x!!.toInt() }
 }
 
 // 0 valueOf
@@ -17,6 +17,6 @@ fun bar() {
 // 1 I2L
 // 2 L2I
 // 2 I2S
-// 2 I2B
+// 1 I2B
 // 1 I2D
 // 1 D2I
